textEntry/commandConsole classes move to shapeCore. command console tests removed. gameloop handle input added.
